requirements.txt
discord.py
components
fuzzywuzzy~=0.18.0
requests
setuptools
wheel
kbbi
python-dotenv
qrcode